requirements.txt
Django==4.0.2
djangorestframework
psycopg2-binary==2.9.3